Make prepared requests reusable
`PreparedRequest::send` currently consumes the request
even though this is not really necessary. This changes
the signature so that a single prepared request can be
reused multiple times.

To make this possible the host header has to be updated
each time, but this actually saves one clone of the URL
with the cost of delaying some of the errors triggered
by `PreparedRequest::set_host`.