Limit buffers::read_line method to avoid unchecked buffer allocations
This makes buffers::read_line stricter in two ways: We limit the amount
of data read and hence the amount of buffers allocated for a single line.
And we actually expect to read a line ending instead of potentially just
hitting EOF as we really always parse full lines in HTTP.

For now this mainly applies to the status and header lines as well as
the chunk sizes for chunked transfer encoding.