Do not try to resolve bracketed IP addresses
URL can contain IP addresses within brackets like "http://[::1]" which should
not be resolve as a domain name but rather used as-is as an IP address.

This change therefore makes use of Url::host instead of Url::host_str to determine
the address to connect with whereas Url::host_str continues to be used for TLS
authentication as this is always based on domain names.